+++ K2D2Project/InfosPages/FindSecrets.cs
@@ -1,7 +1,9 @@
+using AmplifyImpostors;
 using K2D2.KSPService;
 using KSP.Sim;
 using KTools.UI;
 using UnityEngine;
+using BepInEx.Logging;
 
 namespace K2D2.Controller;
 
@@ -22,9 +24,79 @@ public class FindSecrets : BaseController
 
     }
 
+    /*
+
+    Ok trouvé. pour la lune, tout est dans un GameObject nommé Celestial.Mun.Simulation/
+
+    on peux tenter de se baser sur le nom du body en cours pour récupérer cet objet.
+
+    ce go contients un fils nommé objects qui contiens la liste de GameObjects de la planète.
+
+    ils contiennent tous un PQSObject
+
+    */
+
+    public ManualLogSource logger = BepInEx.Logging.Logger.CreateLogSource("K2D2.LandingController");
+
+    List<GameObject> objects = new List<GameObject>();
+
+    void find()
+    {
+        objects.Clear();
+
+        string body_name = current_vessel.VesselComponent.mainBody.Name;
+        last_body_name = body_name;
+        logger.LogMessage($"{body_name} !");
+        string main_go_name = $"Celestial.{body_name}.Simulation";
+        logger.LogMessage($"search for '{main_go_name}'");
+        //GameObject main_go = GameObject.Find(main_go_name);
+        var celestian = Resources.FindObjectsOfTypeAll<GameObject>().Where(obj => obj.name == main_go_name).ToList();
+        if (celestian == null || celestian.Count() == 0)
+        {
+            logger.LogMessage($"error finding '{main_go_name}'");
+            return;
+        }
+
+        // Drawing.
+
+        GameObject main_go = celestian[0];
+        main_go = main_go.transform.FindChild("objects").gameObject;
+
+        foreach(Transform child in main_go.transform)
+        {
+            objects.Add(child.gameObject);
+        }
+    }
+
+    string last_body_name = "";
+
+    Vector2 scrollPos = Vector2.zero;
+
     public override void onGUI()
     {
-        var body = current_vessel.VesselComponent.mainBody;
+        if (UI_Tools.BigButton("find"))
+        {
+            find();
+        }
+
+        if (last_body_name != current_vessel.VesselComponent.mainBody.Name)
+        {
+            objects.Clear();
+        }
+
+        if (objects != null && objects.Count() > 0)
+        {
+            scrollPos = UI_Tools.BeginScrollView(scrollPos, 200);
+
+            foreach(GameObject obj in objects)
+            {
+                UI_Tools.Console(obj.name);
+            }
+
+            GUILayout.EndScrollView();
+        }
+
+        /*var body = current_vessel.VesselComponent.mainBody;
         var transforms = body.transform.children;
         foreach (var child in transforms)
         {
@@ -44,7 +116,7 @@ public class FindSecrets : BaseController
 
             GUILayout.EndHorizontal();
 
-        }
+        }*/
 
 
     }